feat: reward build wip
