+++ src/pages/projectView/projectCreatorViews/sections/contributors/atoms/contributionViewAtom.ts
@@ -1,0 +1,11 @@
+import { atom } from 'jotai'
+
+export enum ContributionView {
+  rewards = 'rewards',
+  pending = 'pending',
+  accounts = 'accounts',
+}
+
+export const contributionViewAtom = atom<ContributionView>(
+  ContributionView.rewards,
+)
