+++ src/pages/projectView/projectCreatorViews/sections/contributors/components/RewardListItem.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const RewardListItem = () => {
+  return (
+    <
+  )
+}
