+++ src/styles/index.ts
@@ -2,3 +2,4 @@ export * from './animations'
 export * from './colors'
 export * from './common'
 export * from './fonts'
+export * from './reponsiveValues'
