+++ src/styles/reponsiveValues.ts
@@ -1,0 +1,1 @@
+export const standardPadding = { base: '10px', lg: '20px' }
